+++ lib/io/dolphin/move/move_notification.dart
@@ -1,0 +1,11 @@
+class MoveNotification {
+  String channelId;
+  String channelName;
+  String channelDescription;
+  String contentTitle;
+  String contentText;
+  String imageName;
+
+  MoveNotification(this.channelId, this.channelName, this.channelDescription,
+      this.contentTitle, this.contentText, this.imageName);
+}